refactor(renderer): modularize renderer into separate modules
- Split renderer.rs into organized module structure:
  - interface.rs: TemplateRenderer trait definition
  - minijinja.rs: MiniJinja implementation
  - filters.rs: Custom filters and string manipulation functions
  - mod.rs: Module organization with re-exports
- Add convenience function new_renderer() for creating renderers
- Update template module to use new convenience function
- Maintain backward compatibility with existing public API
- Follow same pattern as cli, config, loader, prompt, and template modules

All tests pass and functionality remains unchanged.
